+++ src/main/java/com/prgrms/coretime/common/ErrorCode.java
@@ -30,7 +30,17 @@ public enum ErrorCode {
   INVALID_FRIEND_REQUEST_TARGET(400, "F001", "올바르지 않은 친구 상대입니다."),
   DUPLICATE_FRIEND_REQUEST(400, "F002", "이미 완료된 친구 요청입니다."),
   FRIEND_ALREADY_EXISTS(400, "F003", "이미 등록된 친구입니다."),
-  FRIEND_NOT_FOUND(404, "F004", "해당 friend 리소스를 찾을 수 없습니다.");
+  FRIEND_NOT_FOUND(404, "F004", "해당 friend 리소스를 찾을 수 없습니다."),
+
+
+  /**
+   *  Post Domain
+   * */
+  POST_NOT_FOUND(400, "P001", "해당 Post 리소스를 찾을 수 없습니다."),
+  BOARD_NOT_FOUND(400, "B001", "해당 Board 리소스를 찾을 수 없습니다."),
+  PHOTO_NOT_FOUND(400, "PH001", "해당 Photo 리소스를 찾을 수 없습니다."),
+  POST_LIKE_NOT_FOUND(400, "PL001", "해당 User 와 Post 의 좋아요가 존재하지 않습니다."),
+  POST_LIKE_ALREADY_EXISTS(400, "PL002", "해당 User 와 Post 의 좋아요가 이미 존재합니다.");
 
   private final int status;
   private final String code;
